Subject: [PATCH 2/5] x86emu changes from slof-JX-1.0.7-4
implemented bswap opcodes, some tracing fixes, small bugfixes

Acked-by: Myles Watson <mylesgw@gmail.com>


git-svn-id: svn://coreboot.org/repository/coreboot-v3@1074 f3766cd6-281f-0410-b1cd-43a5c92072e9

+++ util/x86emu/x86emu/sys.c
@@ -405,6 +405,6 @@ void X86EMU_prepareForInt(int num)
 
 void X86EMU_setMemBase(void *base, size_t size)
 {
-	M.mem_base = (int) base;
+	M.mem_base = (unsigned long) base;
 	M.mem_size = size;
 }
